no need to store azimuth in double

+++ src/com/incidentlocator/client/IncidentLocatorInterface.java
@@ -10,5 +10,5 @@ import android.location.Location;
 
 public interface IncidentLocatorInterface {
     public void updateLocation(Location loc);
-    public void updateDirection(double azimuth);
+    public void updateDirection(float azimuth);
 }
